doomfront: Update `chumsky`
Remove `must_use` meta from functions returning `impl Parser`, since implementors are now marked with it upstream.

This new version has an interesting new option for error recovery, but existing parsers are unchanged; I'm going to try writing a DECORATE parser in a new way to see what performance looks like, not just for as a price for better error handling but also in terms of Chumsky as a whole.

Unrelatedly, create a `RawParseTree` struct for multithreading parsing, since `rowan::SyntaxNode` is `!Send` and `!Sync`.