DoomFront: start work on DECORATE frontend
I'm not intending to write new parsers until Chumsky's zero-copy overhaul but I can get other work done in the meantime.

This is just some basic parts; syntax tags I'm certain of and a skeleton of an AST root node. This whole language is going to be complex so each commit will just be a moderately-sized chunk.